refactor: use `merge_linters!` macro
