+++ harper-core/src/linting/lint_group.rs
@@ -66,7 +66,6 @@ use super::no_match_for::NoMatchFor;
 use super::nobody::Nobody;
 use super::nominal_wants::NominalWants;
 use super::noun_countability::NounCountability;
-use super::noun_instead_of_verb::NounInsteadOfVerb;
 use super::number_suffix_capitalization::NumberSuffixCapitalization;
 use super::of_course::OfCourse;
 use super::on_floor::OnFloor;
@@ -103,7 +102,6 @@ use super::throw_rubbish::ThrowRubbish;
 use super::touristic::Touristic;
 use super::unclosed_quotes::UnclosedQuotes;
 use super::use_genitive::UseGenitive;
-use super::verb_instead_of_noun::VerbInsteadOfNoun;
 use super::very_unique::VeryUnique;
 use super::was_aloud::WasAloud;
 use super::way_too_adjective::WayTooAdjective;
@@ -444,7 +442,6 @@ impl LintGroup {
         insert_struct_rule!(NoOxfordComma, false);
         insert_expr_rule!(Nobody, true);
         insert_expr_rule!(NounCountability, true);
-        insert_expr_rule!(NounInsteadOfVerb, true);
         insert_struct_rule!(NumberSuffixCapitalization, true);
         insert_struct_rule!(OfCourse, true);
         insert_expr_rule!(OnFloor, true);
@@ -476,7 +473,6 @@ impl LintGroup {
         insert_expr_rule!(Touristic, true);
         insert_struct_rule!(UnclosedQuotes, true);
         insert_expr_rule!(UseGenitive, false);
-        insert_expr_rule!(VerbInsteadOfNoun, true);
         insert_expr_rule!(VeryUnique, true);
         insert_expr_rule!(WasAloud, true);
         insert_expr_rule!(Whereas, true);
