+++ harper-core/src/linting/mod.rs
@@ -72,7 +72,7 @@ mod no_oxford_comma;
 mod nobody;
 mod nominal_wants;
 mod noun_countability;
-mod noun_instead_of_verb;
+mod noun_verb_confusion;
 mod number_suffix_capitalization;
 mod of_course;
 pub mod on_floor;
@@ -115,7 +115,6 @@ mod throw_rubbish;
 mod touristic;
 mod unclosed_quotes;
 mod use_genitive;
-mod verb_instead_of_noun;
 mod very_unique;
 mod was_aloud;
 mod way_too_adjective;
@@ -123,8 +122,6 @@ mod whereas;
 mod widely_accepted;
 mod win_prize;
 mod wordpress_dotcom;
-
-pub mod noun_verb_pairs;
 
 pub use a_part::APart;
 pub use adjective_double_degree::AdjectiveDoubleDegree;
@@ -187,7 +184,7 @@ pub use no_match_for::NoMatchFor;
 pub use no_oxford_comma::NoOxfordComma;
 pub use nobody::Nobody;
 pub use noun_countability::NounCountability;
-pub use noun_instead_of_verb::NounInsteadOfVerb;
+pub use noun_verb_confusion::NounVerbConfusion;
 pub use number_suffix_capitalization::NumberSuffixCapitalization;
 pub use of_course::OfCourse;
 pub use on_floor::OnFloor;
@@ -225,7 +222,6 @@ pub use throw_rubbish::ThrowRubbish;
 pub use touristic::Touristic;
 pub use unclosed_quotes::UnclosedQuotes;
 pub use use_genitive::UseGenitive;
-pub use verb_instead_of_noun::VerbInsteadOfNoun;
 pub use very_unique::VeryUnique;
 pub use was_aloud::WasAloud;
 pub use way_too_adjective::WayTooAdjective;
