--- harper-core/src/linting/noun_verb_pairs.rs
@@ -1,11 +1,0 @@
-/// Common noun-verb pairs that are often confused
-/// See also [`NounInsteadOfVerb`]
-pub const NOUN_VERB_PAIRS: &[(&str, &str)] = &[
-    ("advice", "advise"),
-    ("belief", "believe"),
-    ("breath", "breathe"),
-    ("effect", "affect"), // "Effect" is also a verb meaning "to bring about". "Affect" is a noun in psychology.
-    ("intent", "intend"),
-    // ("proof", "prove"),  // "Proof" is also a verb, a synonym of "proofread".
-    // Add more pairs here as needed
-];
